Image_Recognition_v1.2
The documentation has been improved.
Python files have become more responsive and informative to the console, with displaying the digit number while creating its data.
The model size has increased and now accomodate all the digits and given digits, except {0}. This would be provided in a later update.
The model has become less accurate the accuracy dropped from 98.72% to 94.82%, owing to the increased size of training data.
The model is now saved as 100 iterations as default. This was the demand of the increasing size of data and decreasing accuracy rate.